Always use the scope when resolving expressions.
This is 50% of the fix that I need to do to be realy kosher in terms of
mvvm. The other one is to make sure that scope themselves are resolved
late. I cannot point to objects.

This makes me think that I like a knockoutish system even better now
(and that perhaps I should have followed this model). Anyways...