Support mvvm everywhere.
Until now, mvvm was optional. I made it automatically required. This kinda
sucks in a way (because the expression evaluation thing is relatively big)
but fits the idea that the main focus if the tool is "create client side
applications withoug programming" and mvvm is a key element of that.