Support defaultRoute and model.
Support the defaultRoute and model data to be defined in the editor.
This is still a bit weird (perhaps not), just like the whole question of
setting up the topmost thing in the context of the page: where does the
data come from if there is no outer component...
So here we setup de default route (in a quite ugly way, simply a hash)
and the default model (as a json thing). All this is a starting point for
further reflexions...